refactor: init logic
